+++ projects/poly/code/avg1.py
@@ -2,7 +2,7 @@ def avg(x):
     s = sum(x)
     l = len(x)
 
-    return s/l
+    return int(s/l)
 
 print(avg([1, 2, 1, 1, 1, 3, 1]))
